+++ src/test/kotlin/lotto/ApplicationTest.kt
@@ -7,6 +7,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
 class ApplicationTest : NsTest() {
+
     @Test
     fun `기능 테스트`() {
         assertRandomUniqueNumbersInRangeTest(
